+++ src/check/eval.rs
@@ -222,7 +222,7 @@ fn evaluate_variants(mut tcs: TCS, variants: Vec<LabAbs>) -> (Variants, TCS) {
 fn expand_global(tcs: TCS, expr: Val) -> (Val, TCS) {
     let val = expr.map_neutral(&mut |neut| match neut {
         Neutral::Ref(index) => tcs.glob_val(index).ast.clone(),
-        Neutral::Meta(mi) => match &tcs.meta_solutions()[mi.0] {
+        Neutral::Meta(mi) => match &tcs.meta_context.solutions()[mi.0] {
             MetaSolution::Solved(val) => *val.clone(),
             MetaSolution::Unsolved => panic!("Cannot eval unsolved meta: {:?}", mi),
             MetaSolution::Inlined => unreachable!(),